fix cy.wrap
